Refactoring to adopt SolidJS for all apps
Updated esbuild to 0.17, and made `target` and `platform` configurations explicit.
I also eliminated the "top-level eval" warnings from esbuild, which had some
effect on the bundling.  Finally, I am generating `meta.json` with every prod build.

I refactored `ui/viewer` as `viewer/react`, and moved corresponding files from
`app/classic` into `viewer/solid`.
