+++ .gitignore
@@ -74,3 +74,7 @@ online/public/classic/icons/
 online/src/revision.js
 
 jsweet-branches/
+
+online/meta.json
+
+online/public/modules/
