+++ online/scripts/esbuild-config.mjs
@@ -48,7 +48,7 @@ export const esbuildConfig = {
   // web component, not used by apps
     'vzome-viewer'        : 'src/wc/index.js',
   // client rendering code, dynamically imported for fast time-to-first-render
-    'vzome-viewer-dynamic': 'src/ui/viewer/index.jsx',
+    'vzome-viewer-dynamic': 'src/viewer/react/index.jsx',
   // Worker entry point, only used as a module worker (which still breaks Firefox)
     'vzome-worker-static' : 'src/worker/vzome-worker-static.js',
   // Legacy code, dynamically loaded as needed by the worker to parse vZome files or edit designs
@@ -58,6 +58,8 @@ export const esbuildConfig = {
   splitting: true,
   loader: { '.vef': 'dataurl' },
   format: 'esm',
+  target: 'es2022',
+  platform: 'browser',
   plugins: [ solidPlugin(
     {
       solid: {
