+++ online/src/app/buildplane/index.jsx
@@ -9,7 +9,7 @@ import Link from '@material-ui/core/Link'
 
 import { VZomeAppBar } from '../components/appbar.jsx'
 import { BuildPlaneTool } from './buildplane.jsx'
-import { DesignViewer, WorkerContext } from '../../ui/viewer/index.jsx'
+import { DesignViewer, WorkerContext } from '../../viewer/react/index.jsx'
 import { createWorker } from "../../workerClient/client.js";
 
 const worker = createWorker();
