+++ online/src/app/buildplane/planes.js
@@ -1,4 +1,3 @@
-import shapes from "../../worker/legacy/resources/com/vzome/core/parts";
 
 export const doStrutPreview= ( position ) =>
 {
